Refactoring HttpClient implementations
